fix(deps): update dependency cache-manager to v5

+++ packages/cache-manager-ioredis/src/store.ts
@@ -1,13 +1,12 @@
 /* eslint-disable prefer-rest-params */
-import { CacheManager, CacheManagerSetOptions, chunk, isNullOrUndefined } from "@anchan828/nest-cache-common";
-import { CacheStore, CacheStoreFactory, LiteralObject } from "@nestjs/common";
+import { CacheManager, chunk, isNullOrUndefined } from "@anchan828/nest-cache-common";
 import Redis from "ioredis";
 import { pack, unpack } from "msgpackr";
 import { AsyncLocalStorageService } from "./async-local-storage.service";
 import { CACHE_STORE_NAME } from "./constants";
 import { RedisStoreArgs } from "./store.interface";
 export class RedisStore implements CacheManager {
-  private readonly client: Redis;
+  readonly store: Redis;
 
   public readonly name: string = CACHE_STORE_NAME;
 
@@ -15,26 +14,33 @@ export class RedisStore implements CacheManager {
 
   constructor(private readonly args: RedisStoreArgs) {
     this.asyncLocalStorage = new AsyncLocalStorageService(args.asyncLocalStorage);
-    this.client = args.client ?? new Redis(args);
+    this.store = args.client ?? new Redis(args);
   }
 
-  public async set<T = any>(key: string, value: T, options?: CacheManagerSetOptions): Promise<void> {
+  public async ttl(key: string): Promise<number> {
+    if (typeof key !== "string") {
+      return 0;
+    }
+    return this.store.ttl(key);
+  }
+
+  public async set<T = any>(key: string, value: T, ttl?: number): Promise<void> {
+    if (typeof key !== "string") {
+      return;
+    }
+
     if (isNullOrUndefined(value)) {
       return;
     }
 
-    let ttl: number | undefined;
-
-    if (options && !isNullOrUndefined(options.ttl)) {
-      ttl = options.ttl;
-    } else if (!isNullOrUndefined(this.args.ttl)) {
+    if (isNullOrUndefined(ttl) && !isNullOrUndefined(this.args.ttl)) {
       ttl = this.args.ttl;
     }
 
     if (!isNullOrUndefined(ttl) && ttl !== 0 && ttl !== -1) {
-      this.client.setex(key, ttl, pack(value));
+      this.store.setex(key, ttl, pack(value));
     } else if (ttl !== 0) {
-      this.client.set(key, pack(value));
+      this.store.set(key, pack(value));
     } else {
       return;
     }
@@ -54,7 +60,7 @@ export class RedisStore implements CacheManager {
       return result;
     }
 
-    const rawResult = await this.client.getBuffer(key);
+    const rawResult = await this.store.getBuffer(key);
 
     if (isNullOrUndefined(rawResult)) {
       return;
@@ -67,14 +73,13 @@ export class RedisStore implements CacheManager {
     return result;
   }
 
-  public async del(...keys: string[]): Promise<void> {
-    for (const deleteKeys of chunk(keys, 2000)) {
-      await this.client.del(...deleteKeys);
-      for (const key of deleteKeys) {
-        this.asyncLocalStorage.delete(key);
-        await this.args.hooks?.delete?.(key, undefined);
-      }
+  public async del(key: string): Promise<void> {
+    if (typeof key !== "string") {
+      return;
     }
+    await this.store.del(key);
+    this.asyncLocalStorage.delete(key);
+    await this.args.hooks?.delete?.(key, undefined);
   }
 
   public async keys(pattern?: string): Promise<string[]> {
@@ -83,7 +88,7 @@ export class RedisStore implements CacheManager {
       pattern = "*";
     }
 
-    for await (const keys of this.client.scanStream({ match: pattern, count: 100 })) {
+    for await (const keys of this.store.scanStream({ match: pattern, count: 100 })) {
       results.push(...keys);
     }
 
@@ -94,20 +99,16 @@ export class RedisStore implements CacheManager {
     this.asyncLocalStorage.clear();
     const keys = await this.keys();
     if (keys.length !== 0) {
-      await this.del(...keys.map((key) => key.replace(new RegExp(`^${this.args.keyPrefix}`), "")));
+      await this.mdel(...keys.map((key) => key.replace(new RegExp(`^${this.args.keyPrefix}`), "")));
     }
   }
 
-  public async mget<T>(...keysOrOptions: string[]): Promise<Array<T | undefined>> {
-    const keys = keysOrOptions.filter((x): x is string => typeof x === "string") as string[];
+  public async mget<T>(...keys: string[]): Promise<Array<T | undefined>> {
+    const map = new Map<string, T | undefined>(
+      keys.filter((key) => typeof key === "string").map((key) => [key, undefined]),
+    );
 
-    const map = new Map<string, T | undefined>(keys.map((key) => [key, undefined]));
-
-    for (const key of keys) {
-      if (typeof key !== "string") {
-        continue;
-      }
-
+    for (const key of map.keys()) {
       const result = this.asyncLocalStorage.get(key);
       if (!isNullOrUndefined(result)) {
         map.set(key, result);
@@ -120,7 +121,7 @@ export class RedisStore implements CacheManager {
       const results: Array<Buffer | null> = [];
 
       for (const keys of chunk(notFoundKeys, 2000)) {
-        results.push(...((await this.client.mgetBuffer(...keys)) as Array<Buffer | null>));
+        results.push(...((await this.store.mgetBuffer(...keys)) as Array<Buffer | null>));
       }
 
       for (let index = 0; index < notFoundKeys.length; index++) {
@@ -139,51 +140,54 @@ export class RedisStore implements CacheManager {
     return [...map.values()];
   }
 
-  public async mset<T>(...keyOrValues: Array<string | T | CacheManagerSetOptions>): Promise<void> {
-    let options: CacheManagerSetOptions | undefined;
-
-    if (keyOrValues.length % 2 > 0 && this.isObject(keyOrValues[keyOrValues.length - 1])) {
-      options = keyOrValues[keyOrValues.length - 1] as CacheManagerSetOptions;
+  public async mset<T>(keyOrValues: [string, T][], ttl?: number): Promise<void> {
+    if (isNullOrUndefined(ttl) && !isNullOrUndefined(this.args.ttl)) {
+      ttl = this.args.ttl;
     }
 
-    for (let i = 0; i < keyOrValues.length; i += 2) {
-      if (keyOrValues.length !== i + 1) {
-        const key = keyOrValues[i] as string;
-        const value = keyOrValues[i + 1];
-        if (typeof key !== "string") {
-          continue;
-        }
+    if (ttl === 0) {
+      return;
+    }
 
-        if (isNullOrUndefined(value)) {
-          continue;
-        }
+    for (const [key, value] of keyOrValues) {
+      if (typeof key !== "string") {
+        continue;
+      }
 
-        let ttl: number | undefined;
+      if (isNullOrUndefined(value)) {
+        continue;
+      }
 
-        if (options && !isNullOrUndefined(options.ttl)) {
-          ttl = options.ttl;
-        } else if (!isNullOrUndefined(this.args.ttl)) {
-          ttl = this.args.ttl;
-        }
+      if (!isNullOrUndefined(ttl) && ttl !== -1) {
+        this.store.setex(key, ttl, pack(value));
+      } else {
+        this.store.set(key, pack(value));
+      }
 
-        if (ttl === 0) {
-          continue;
-        }
+      await this.args.hooks?.set?.(key, undefined, value, ttl);
 
-        if (!isNullOrUndefined(ttl) && ttl !== -1) {
-          this.client.setex(key, ttl, pack(value));
-        } else {
-          this.client.set(key, pack(value));
-        }
+      this.asyncLocalStorage.set(key, value);
+    }
+  }
 
-        await this.args.hooks?.set?.(key, undefined, value, ttl);
-
-        this.asyncLocalStorage.set(key, value);
+  public async mdel(...keys: string[]): Promise<void> {
+    for (const deleteKeys of chunk(
+      keys.filter((key) => typeof key === "string"),
+      2000,
+    )) {
+      await this.store.del(...deleteKeys);
+      for (const key of deleteKeys) {
+        this.asyncLocalStorage.delete(key);
+        await this.args.hooks?.delete?.(key, undefined);
       }
     }
   }
 
   public async hget<T>(key: string, field: string): Promise<T | undefined> {
+    if (typeof key !== "string" || typeof field !== "string") {
+      return;
+    }
+
     const asyncLocalStorageKey = `${key}:h:${field}`;
 
     let result: T | null | undefined = this.asyncLocalStorage.get<T>(asyncLocalStorageKey);
@@ -192,7 +196,7 @@ export class RedisStore implements CacheManager {
       return result;
     }
 
-    const rawResult = await this.client.hgetBuffer(key, field);
+    const rawResult = await this.store.hgetBuffer(key, field);
 
     if (isNullOrUndefined(rawResult)) {
       return;
@@ -207,13 +211,13 @@ export class RedisStore implements CacheManager {
   }
 
   public async hset<T>(key: string, field: string, value: T): Promise<void> {
-    if (isNullOrUndefined(value)) {
+    if (typeof key !== "string" || typeof field !== "string" || isNullOrUndefined(value)) {
       return;
     }
 
     const asyncLocalStorageKey = `${key}:h:${field}`;
 
-    this.client.hset(key, field, pack(value));
+    this.store.hset(key, field, pack(value));
 
     this.asyncLocalStorage.set(asyncLocalStorageKey, value);
 
@@ -221,8 +225,14 @@ export class RedisStore implements CacheManager {
   }
 
   public async hdel(key: string, ...fields: string[]): Promise<void> {
-    for (const deleteFields of chunk(fields, 2000)) {
-      await this.client.hdel(key, ...deleteFields);
+    if (typeof key !== "string") {
+      return;
+    }
+    for (const deleteFields of chunk(
+      fields.filter((field) => typeof field === "string"),
+      2000,
+    )) {
+      await this.store.hdel(key, ...deleteFields);
       for (const field of deleteFields) {
         this.asyncLocalStorage.delete(`${key}:h:${field}`);
         await this.args.hooks?.delete?.(key, field);
@@ -231,7 +241,10 @@ export class RedisStore implements CacheManager {
   }
 
   public async hgetall(key: string): Promise<Record<string, any>> {
-    const rawResults = await this.client.hgetallBuffer(key);
+    if (typeof key !== "string") {
+      return {};
+    }
+    const rawResults = await this.store.hgetallBuffer(key);
     const results: Record<string, any> = {};
 
     for (const [key, value] of Object.entries(rawResults)) {
@@ -242,19 +255,18 @@ export class RedisStore implements CacheManager {
   }
 
   public async hkeys(key: string): Promise<string[]> {
-    return await this.client.hkeys(key);
+    if (typeof key !== "string") {
+      return [];
+    }
+    return await this.store.hkeys(key);
   }
 
   public async close(): Promise<void> {
-    await this.client.quit();
+    await this.store.quit();
     this.asyncLocalStorage.clear();
-  }
-
-  private isObject(value: any): value is Object {
-    return value instanceof Object && value.constructor === Object;
   }
 }
 
-export const redisStore: CacheStoreFactory = {
-  create: (args: LiteralObject): CacheStore => new RedisStore(args),
-};
+export function redisStore(args: RedisStoreArgs): RedisStore {
+  return new RedisStore(args);
+}
